+++ docs/app/breeze/layout.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from "react";
 import { DocsLayout } from "fumadocs-ui/layouts/notebook";
 import { breezeOptions } from "../layout.config";
-import { RootProvider } from "fumadocs-ui/provider";
+import { RootProvider } from "fumadocs-ui/provider/next";
 import DefaultSearchDialog from "@/components/search/search";
 import { TAGS } from "@/app/api/search/constants";
 import { MotionProvider } from "@/components/MotionProvider";