+++ docs/examples/react/image-frame/overlay.tsx
@@ -10,6 +10,7 @@ import {
   VStack,
   Text,
 } from "@seed-design/react";
+import { ContentPlaceholder } from "seed-design/ui/content-placeholder";
 import { useState } from "react";
 
 export default function ImageFrameOverlayExample() {
@@ -25,6 +26,7 @@ export default function ImageFrameOverlayExample() {
           src="https://images.unsplash.com/photo-1535025183041-0991a977e25b?w=300&dpr=2&q=80"
           alt="Landscape with badge overlay"
           style={{ width: 120 }}
+          fallback={<ContentPlaceholder type="buySell" />}
         >
           <ImageFrameFloater placement="bottom-end">
             <ImageFrameBadge tone="brand" variant="solid">
@@ -45,6 +47,7 @@ export default function ImageFrameOverlayExample() {
           src="https://images.unsplash.com/photo-1535025183041-0991a977e25b?w=300&dpr=2&q=80"
           alt="Landscape with icon overlay"
           style={{ width: 120 }}
+          fallback={<ContentPlaceholder type="commerce" />}
         >
           <ImageFrameFloater placement="bottom-end">
             <ImageFrameIcon svg={<IconCarrotFill />} />
@@ -63,6 +66,7 @@ export default function ImageFrameOverlayExample() {
           src="https://images.unsplash.com/photo-1535025183041-0991a977e25b?w=300&dpr=2&q=80"
           alt="Landscape with indicator overlay"
           style={{ width: 120 }}
+          fallback={<ContentPlaceholder type="food" />}
         >
           <ImageFrameFloater placement="bottom-end">
             <ImageFrameIndicator>+9</ImageFrameIndicator>
@@ -81,6 +85,7 @@ export default function ImageFrameOverlayExample() {
           src="https://images.unsplash.com/photo-1535025183041-0991a977e25b?w=300&dpr=2&q=80"
           alt="Landscape with reaction button overlay"
           style={{ width: 120 }}
+          fallback={<ContentPlaceholder type="car" />}
         >
           <ImageFrameFloater placement="bottom-end">
             <ImageFrameReactionButton