+++ packages/react/src/components/ContentPlaceholder/ContentPlaceholder.namespace.ts
@@ -1,0 +1,6 @@
+export {
+  ContentPlaceholderRoot as Root,
+  ContentPlaceholderAsset as Asset,
+  type ContentPlaceholderRootProps as RootProps,
+  type ContentPlaceholderAssetProps as AssetProps,
+} from "./ContentPlaceholder";