+++ packages/react/src/components/index.ts
@@ -14,6 +14,7 @@ export * from "./Chip";
 export * from "./ChipTabs";
 export * from "./Columns";
 export * from "./ConsistentWidth";
+export * from "./ContentPlaceholder";
 export * from "./ContextualFloatingButton";
 export * from "./ControlChip";
 export * from "./Count";